requirements.txt
# ===========================================================================
# Packages essential to this app. Needed by anyone who wants to use this app.
# ===========================================================================
django
django-hvad
django-libs
Pillow
django-filer
django-cms
django-localized-names
